requirements-dev.txt
datetime
load_dotenv
bs4
aiohttp
pytest
pytest-asyncio